11. Нейросеть для дополнения текстов/requirements..txt
torch>=2.0.0
pandas>=1.5.0
scikit-learn>=1.2.0
transformers>=4.30.0
evaluate>=0.4.0
numpy>=1.24.0
pyyaml>=6.0
rouge-score>=0.1.2
tqdm>=4.65.0